Revert "Revert "We can remove NioSocketChannel at least... possibly others as well.  These were likely required in previous versions of""
This reverts commit 3cdf5fee367c863df86b6a4f734cd48bc65a610f.

Hmmm, a revert of a revert.. those are always tasty.

Original revert was motivated by the appearance of https://github.com/oracle/helidon/pull/1722 in the Helidon test case.  Thing is
this can't be reproduced in non-Helidon contexts with the same GraalVM version.  And it's definitely the case that a no-arg
constructor is present (see
https://github.com/netty/netty/blob/netty-4.1.60.Final/transport/src/main/java/io/netty/channel/socket/nio/NioSocketChannel.java).
So... this looks to be something unique to Helidon.